Subset columns of a linelist object

This function was deprecated to ensure full compatibility with the default dplyr::select() methods. The tag selection feature is now possible via the has_tag() selection helper.

Usage

# S3 method for class 'linelist'
select(.data, ..., tags)

Arguments

.data

a linelist object

...

the variables to select, using dplyr compatible syntax

tags

It is now recommended to leverage the has_tag() selection helper rather than this argument.

Value

The function returns a linelist with selected columns.

See also