Subsetting of linelist objects

The [] and [[]] operators for linelist objects behaves like for regular data.frame or tibble, but check that tagged variables are not lost, and takes the appropriate action if this is the case (warning, error, or ignore, depending on the general option set via lost_tags_action()) .

Usage

# S3 method for class 'linelist'
x[i, j, drop = FALSE]

# S3 method for class 'linelist'
x[i, j] <- value

# S3 method for class 'linelist'
x[[i, j]] <- value

# S3 method for class 'linelist'
x$name <- value

Arguments

x

a linelist object

i

a vector of integer or logical to subset the rows of the linelist

j

a vector of character, integer, or logical to subset the columns of the linelist

drop

a logical indicating if, when a single column is selected, the data.frame class should be dropped to return a simple vector, in which case the linelist class is lost as well; defaults to FALSE

value

the replacement to be used for the entries identified in x

name

a literal character string or a name (possibly backtick quoted). For extraction, this is normally (see under ‘Environments’) partially matched to the names of the object.

Value

If no drop is happening, a linelist. Otherwise an atomic vector.

See also

Examples

if (require(outbreaks) && require(dplyr) && require(magrittr)) {
  ## create a linelist
  x <- measles_hagelloch_1861 %>%
    make_linelist(
      id = "case_ID",
      date_onset = "date_of_prodrome",
      age = "age",
      gender = "gender"
    ) %>%
    mutate(result = if_else(is.na(date_of_death), "survived", "died")) %>%
    set_tags(outcome = "result") %>%
    rename(identifier = case_ID)
  x

  ## dangerous removal of a tagged column setting it to NULL issues a warning
  x[, 1] <- NULL
  x

  x[[2]] <- NULL
  x

  x$age <- NULL
  x
}
#> Warning: The following tags have lost their variable:
#>  id:identifier
#> Warning: The following tags have lost their variable:
#>  date_onset:date_of_prodrome
#> Warning: The following tags have lost their variable:
#>  age:age
#> 
#> // linelist object
#>     infector date_of_rash date_of_death gender family_ID class complications
#> 1         45   1861-11-25          <NA>      f        41     1           yes
#> 2         45   1861-11-27          <NA>      f        41     1           yes
#> 3        172   1861-12-02          <NA>      f        41     0           yes
#> 4        180   1861-11-28          <NA>      m        61     2           yes
#> 5         45   1861-11-27          <NA>      f        42     1           yes
#> 6        180   1861-11-29          <NA>      m        42     2           yes
#> 7         42   1861-11-28          <NA>      m        26     0           yes
#> 8         45   1861-11-26          <NA>      m        44     1           yes
#> 9        182   1861-11-30          <NA>      m        44     2           yes
#> 10        45   1861-11-25          <NA>      f        29     1           yes
#> 11       182   1861-11-30          <NA>      f        27     2           yes
#> 12        45   1861-11-25          <NA>      f        32     1           yes
#> 13        12   1861-12-05          <NA>      m        32     2           yes
#> 14       181   1861-11-29          <NA>      f        22     2           yes
#> 15        45   1861-11-29          <NA>      m        22     1           yes
#> 16       181   1861-11-25          <NA>      f        43     2           yes
#> 17       181   1861-11-25          <NA>      f        43     2           yes
#> 18       175   1861-11-27          <NA>      f        43     0           yes
#> 19       181   1861-11-24          <NA>      m        11     2           yes
#> 20       181   1861-11-27          <NA>      m        11     2           yes
#> 21       181   1861-11-29          <NA>      f        11     2           yes
#> 22        45   1861-11-25          <NA>      f        35     1           yes
#> 23        45   1861-11-25          <NA>      f        35     1           yes
#> 24        22   1861-12-04          <NA>      m        35     0           yes
#> 25        22   1861-12-04          <NA>      f        35     2           yes
#> 26        45   1861-11-27          <NA>      m        67     1           yes
#> 27        10   1861-12-02    1861-12-14      m        29     0           yes
#> 28       180   1861-11-30          <NA>      f        65     2           yes
#> 29        31   1861-12-02          <NA>      m        15     2           yes
#> 30        45   1861-11-29          <NA>      f        15     2           yes
#> 31        45   1861-11-24          <NA>      f        15     1           yes
#> 32        45   1861-11-24          <NA>      m        15     1           yes
#> 33        45   1861-11-26          <NA>      f         1     1           yes
#> 34       181   1861-11-26          <NA>      m         1     2           yes
#> 35       182   1861-11-30          <NA>      f        10     2           yes
#> 36        34   1861-12-05          <NA>      m        46     2           yes
#> 37       182   1861-12-02          <NA>      m        46     2           yes
#> 38        17   1861-12-05          <NA>   <NA>        54     2           yes
#> 39        45   1861-11-26          <NA>      m         5     1           yes
#> 40        93   1861-12-07          <NA>      m        36     2           yes
#> 41       180   1861-11-29          <NA>      m        59     2           yes
#> 42       178   1861-11-19          <NA>      f        33     0           yes
#> 43        42   1861-11-30          <NA>      m        33     0           yes
#> 44        45   1861-11-24          <NA>      m        45     1           yes
#> 45       184   1861-11-13          <NA>      m        48     1           yes
#> 46        45   1861-11-27          <NA>      m        48     2           yes
#> 47        45   1861-11-27          <NA>      f        48     2           yes
#> 48        10   1861-12-05          <NA>      f        29     0           yes
#> 49        17   1861-12-06          <NA>      m        49     2           yes
#> 50         8   1861-12-06          <NA>      f        44     0           yes
#> 51        31   1861-12-05          <NA>      m        15     0           yes
#> 52        17   1861-12-05    1861-12-12      m        43     2           yes
#> 53        17   1861-12-04          <NA>      m        43     0           yes
#> 54        17   1861-12-05    1861-12-15      m        43     0           yes
#> 55        17   1861-12-05          <NA>      f        43     2           yes
#> 56        45   1861-11-29          <NA>      m        20     1           yes
#> 57        56   1861-12-05          <NA>      m        20     0           yes
#> 58        45   1861-11-26          <NA>      m        24     1           yes
#> 59        58   1861-12-04          <NA>      f        24     2           yes
#> 60        58   1861-12-06    1861-12-12      m        24     0           yes
#> 61       186   1861-12-03          <NA>      f         8     0           yes
#> 62        11   1861-12-06          <NA>      m         8     2           yes
#> 63        19   1861-12-08          <NA>      m         8     2           yes
#> 64        45   1861-11-27          <NA>      m        30     1           yes
#> 65        64   1861-12-06          <NA>      m        30     0           yes
#> 66        64   1861-12-05          <NA>      f        30     0           yes
#> 67        11   1861-12-07          <NA>      f        55     2           yes
#> 68       179   1861-12-02          <NA>      f        16     0           yes
#> 69        54   1861-12-17          <NA>      m        63     0           yes
#> 70       180   1861-12-03          <NA>      m        21     2           yes
#> 71        10   1861-12-05          <NA>      f        29     2           yes
#> 72        12   1861-12-06          <NA>      f        32     2           yes
#> 73       180   1861-12-05          <NA>      f        47     2           yes
#> 74        45   1861-11-26          <NA>      f        23     1           yes
#> 75        74   1861-12-05          <NA>      m        23     2           yes
#> 76         5   1861-12-06          <NA>      m        42     0           yes
#> 77       180   1861-11-27          <NA>      f        18     1           yes
#> 78       181   1861-11-26          <NA>      f        18     2           yes
#> 79       179   1861-12-05          <NA>      f        18     0           yes
#> 80        78   1861-12-05    1861-12-12      m        18     2           yes
#> 81        39   1861-12-04          <NA>      m         5     2           yes
#> 82        45   1861-11-26          <NA>      m        21     1           yes
#> 83        82   1861-12-05          <NA>      f        21     2           yes
#> 84        82   1861-12-04          <NA>      m        21     0           yes
#> 85        44   1861-12-05          <NA>      f        45     0           yes
#> 86         1   1861-12-05          <NA>      f        41     0           yes
#> 87        47   1861-12-07          <NA>      f        60     2           yes
#> 88        47   1861-12-07          <NA>      f        60     2           yes
#> 89        12   1861-12-05          <NA>      f        32     0           yes
#> 90        93   1861-12-05          <NA>      m        36     2           yes
#> 91        93   1861-12-05          <NA>      f        36     0           yes
#> 92        93   1861-12-07          <NA>      f        36     0           yes
#> 93        45   1861-11-26          <NA>      m        36     1           yes
#> 94       183   1861-12-03          <NA>      f        10     0           yes
#> 95        10   1861-12-07          <NA>      m        29     0           yes
#> 96        97   1861-12-04          <NA>      m        28     0           yes
#> 97        45   1861-11-25          <NA>      m        28     1           yes
#> 98        64   1861-12-12    1861-12-17      m        27     0           yes
#> 99        11   1861-12-10          <NA>      m        27     0           yes
#> 100       47   1861-12-07          <NA>      m        69     2           yes
#> 101        7   1861-12-06          <NA>      f        26     0           yes
#> 102       21   1861-12-07          <NA>      m        68     2           yes
#> 103       37   1861-12-07          <NA>   <NA>        54     2           yes
#> 104       58   1861-12-06          <NA>      m        24     0           yes
#> 105       74   1861-12-08          <NA>      m        23     0           yes
#> 106       42   1861-11-26          <NA>      m        34     0           yes
#> 107       19   1861-12-07          <NA>      f        34     2           yes
#> 108      106   1861-12-06          <NA>      m        34     0           yes
#> 109       12   1861-12-07          <NA>      f        32     0           yes
#> 110       18   1861-12-06          <NA>      m        49     0           yes
#> 111       34   1861-12-05          <NA>      m         2     2           yes
#> 112       21   1861-12-08          <NA>      m         2     2           yes
#> 113       31   1861-12-07          <NA>      f        15     0           yes
#> 114       78   1861-12-05          <NA>   <NA>        52     2           yes
#> 115       16   1861-12-07          <NA>      f        66     2           yes
#> 116       45   1861-11-25          <NA>      m        40     1           yes
#> 117      116   1861-12-06          <NA>      m        40     2           yes
#> 118      116   1861-12-06          <NA>      f        40     0           yes
#> 119      116   1861-12-08          <NA>      f        40     0           yes
#> 120        7   1861-12-08          <NA>      m        26     0           yes
#> 121       11   1861-12-08          <NA>      m        50     2           yes
#> 122      188   1861-12-10          <NA>      f        50     0           yes
#> 123        7   1861-12-07    1861-12-13      f        25     0           yes
#> 124        7   1861-12-08          <NA>      m         7     0           yes
#> 125        7   1861-12-09          <NA>      f         7     0           yes
#> 126       37   1861-12-09          <NA>   <NA>        54     0           yes
#> 127      106   1861-12-09    1861-12-15      m        34     0           yes
#> 128        7   1861-12-05          <NA>      m         6     0           yes
#> 129        7   1861-12-07          <NA>      f         6     0           yes
#> 130       56   1861-12-06          <NA>      f        20     2           yes
#> 131       56   1861-12-06          <NA>      m        20     0           yes
#> 132       14   1861-12-07          <NA>      m        22     2           yes
#> 133       18   1861-12-08          <NA>      f        53     0           yes
#> 134       78   1861-12-06          <NA>      m        18     2           yes
#> 135       79   1861-12-07          <NA>      m        18     0           yes
#> 136       17   1861-12-07          <NA>      f        12     2           yes
#> 137       16   1861-12-06          <NA>      f        12     2           yes
#> 138       34   1861-12-05          <NA>   <NA>         3     2           yes
#> 139        4   1861-12-11          <NA>   <NA>         3     2           yes
#> 140        6   1861-12-10          <NA>   <NA>        58     0           yes
#> 141       NA   1862-01-27          <NA>      f        50     0           yes
#> 142      145   1861-12-08    1861-12-09      m        39     0           yes
#> 143      145   1861-12-08          <NA>      f        39     2           yes
#> 144      145   1861-12-08          <NA>      m        39     0           yes
#> 145       45   1861-11-27          <NA>      f        39     1           yes
#> 146      172   1861-12-07          <NA>      f        13     0           yes
#> 147       18   1861-12-07          <NA>      f        13     0           yes
#> 148       14   1861-12-07          <NA>      f        19     2           yes
#> 149       39   1861-12-09    1861-12-15      m        19     0           yes
#> 150      148   1861-12-15          <NA>      m        19     0           yes
#> 151      153   1861-12-11          <NA>      f        37     0           yes
#> 152      153   1861-12-08          <NA>      f        37     0           yes
#> 153       45   1861-11-27          <NA>      f        37     1           yes
#> 154      153   1861-12-04          <NA>      f        37     0           yes
#> 155       73   1861-12-13          <NA>      m        47     2           yes
#> 156       45   1861-11-24          <NA>      m        31     1           yes
#> 157      156   1861-12-07          <NA>      f        31     0           yes
#> 158      156   1861-12-07          <NA>      f        31     0           yes
#> 159       37   1861-12-13          <NA>      f        46     0           yes
#> 160       68   1861-12-13          <NA>      f        16     0           yes
#> 161      148   1861-12-16          <NA>      f        19     0           yes
#> 162      123   1861-12-16          <NA>      f        25     0           yes
#> 163      123   1861-12-16    1861-12-28      f        25     0           yes
#> 164      102   1861-12-15          <NA>      m        14     0           yes
#> 165      102   1861-12-19          <NA>      m        14     0           yes
#> 166      153   1861-12-07          <NA>      m        37     0           yes
#> 167      110   1861-12-18          <NA>      m        49     0           yes
#> 168       98   1861-12-15          <NA>      m        56     0           yes
#> 169      153   1861-12-07          <NA>      m        38     0           yes
#> 170      153   1861-12-10          <NA>      m        38     2           yes
#> 171      169   1861-12-17          <NA>      m        38     0           yes
#> 172      174   1861-11-21          <NA>      f        62     0           yes
#> 173       NA   1861-11-03          <NA>      m         9     0           yes
#> 174       NA   1861-11-08          <NA>      f         9     0           yes
#> 175      173   1861-11-15          <NA>      m         9     0           yes
#> 176      146   1861-12-15          <NA>   <NA>        64     0           yes
#> 177      184   1861-11-11          <NA>      f        17     1           yes
#> 178      184   1861-11-11          <NA>      m        17     0           yes
#> 179      177   1861-11-21          <NA>      f        17     1           yes
#> 180      177   1861-11-21          <NA>      m        17     2           yes
#> 181      184   1861-11-17          <NA>      f         4     2           yes
#> 182      184   1861-11-18    1861-11-18      m         4     2           yes
#> 183      184   1861-11-15          <NA>      m         4     0           yes
#> 184       NA   1861-11-06          <NA>   <NA>        51     2           yes
#> 185       82   1861-12-07          <NA>      m        21     0           yes
#> 186       45   1861-11-26          <NA>   <NA>        57     0           yes
#> 187       82   1861-12-11          <NA>      m        21     0           yes
#> 188      175   1861-11-27          <NA>   <NA>        57     0           yes
#>     x_loc y_loc   result
#> 1   142.5 100.0 survived
#> 2   142.5 100.0 survived
#> 3   142.5 100.0 survived
#> 4   165.0 102.5 survived
#> 5   145.0 120.0 survived
#> 6   145.0 120.0 survived
#> 7   272.5 147.5 survived
#> 8    97.5 155.0 survived
#> 9    97.5 155.0 survived
#> 10  240.0  75.0 survived
#> 11  270.0 135.0 survived
#> 12  195.0  27.5 survived
#> 13  195.0  27.5 survived
#> 14  227.5 185.0 survived
#> 15  227.5 185.0 survived
#> 16  172.5 172.5 survived
#> 17  172.5 172.5 survived
#> 18  172.5 172.5 survived
#> 19  167.5   5.0 survived
#> 20  167.5   5.0 survived
#> 21  167.5   5.0 survived
#> 22  167.5   5.0 survived
#> 23  167.5   5.0 survived
#> 24  167.5   5.0 survived
#> 25  167.5   5.0 survived
#> 26    7.5  37.5 survived
#> 27  240.0  75.0     died
#> 28   15.0  47.5 survived
#> 29  125.0 187.5 survived
#> 30  125.0 187.5 survived
#> 31  125.0 187.5 survived
#> 32  125.0 187.5 survived
#> 33  125.0 187.5 survived
#> 34  125.0 187.5 survived
#> 35  190.0 115.0 survived
#> 36  280.0 192.5 survived
#> 37  280.0 192.5 survived
#> 38  280.0 192.5 survived
#> 39  227.5 217.5 survived
#> 40  162.5  47.5 survived
#> 41  185.0  80.0 survived
#> 42  170.0  17.5 survived
#> 43  170.0  17.5 survived
#> 44  150.0  15.0 survived
#> 45   75.0  55.0 survived
#> 46   75.0  55.0 survived
#> 47   75.0  55.0 survived
#> 48  240.0  75.0 survived
#> 49  175.0 140.0 survived
#> 50   97.5 155.0 survived
#> 51  125.0 187.5 survived
#> 52  172.5 172.5     died
#> 53  172.5 172.5 survived
#> 54  172.5 172.5     died
#> 55  172.5 172.5 survived
#> 56  265.0 225.0 survived
#> 57  265.0 225.0 survived
#> 58  250.0 210.0 survived
#> 59  250.0 210.0 survived
#> 60  250.0 210.0     died
#> 61  270.0 102.5 survived
#> 62  270.0 102.5 survived
#> 63  270.0 102.5 survived
#> 64  237.5  90.0 survived
#> 65  237.5  90.0 survived
#> 66  237.5  90.0 survived
#> 67  267.5 127.5 survived
#> 68  165.0 192.5 survived
#> 69  150.0 165.0 survived
#> 70  205.0 182.5 survived
#> 71  240.0  75.0 survived
#> 72  195.0  27.5 survived
#> 73  182.5  55.0 survived
#> 74  257.5 195.0 survived
#> 75  257.5 195.0 survived
#> 76  145.0 120.0 survived
#> 77  187.5 240.0 survived
#> 78  187.5 240.0 survived
#> 79  187.5 240.0 survived
#> 80  187.5 240.0     died
#> 81  227.5 217.5 survived
#> 82  205.0 182.5 survived
#> 83  205.0 182.5 survived
#> 84  205.0 182.5 survived
#> 85  150.0  15.0 survived
#> 86  142.5 100.0 survived
#> 87   77.5  42.5 survived
#> 88   77.5  42.5 survived
#> 89  195.0  27.5 survived
#> 90  162.5  47.5 survived
#> 91  162.5  47.5 survived
#> 92  162.5  47.5 survived
#> 93  162.5  47.5 survived
#> 94  190.0 115.0 survived
#> 95  240.0  75.0 survived
#> 96  230.0 120.0 survived
#> 97  230.0 120.0 survived
#> 98  270.0 135.0     died
#> 99  270.0 135.0 survived
#> 100 252.5 150.0 survived
#> 101 272.5 147.5 survived
#> 102 257.5 180.0 survived
#> 103 280.0 192.5 survived
#> 104 250.0 210.0 survived
#> 105 257.5 195.0 survived
#> 106 170.0  17.5 survived
#> 107 170.0  17.5 survived
#> 108 170.0  17.5 survived
#> 109 195.0  27.5 survived
#> 110 175.0 140.0 survived
#> 111 142.5 180.0 survived
#> 112 142.5 180.0 survived
#> 113 125.0 187.5 survived
#> 114 200.0 210.0 survived
#> 115  75.0  20.0 survived
#> 116 127.5 147.5 survived
#> 117 127.5 147.5 survived
#> 118 127.5 147.5 survived
#> 119 127.5 147.5 survived
#> 120 272.5 147.5 survived
#> 121 252.5 150.0 survived
#> 122 252.5 150.0 survived
#> 123 280.0 157.5     died
#> 124 280.0 167.5 survived
#> 125 280.0 167.5 survived
#> 126 280.0 192.5 survived
#> 127 170.0  17.5     died
#> 128 240.0 225.0 survived
#> 129 240.0 225.0 survived
#> 130 265.0 225.0 survived
#> 131 265.0 225.0 survived
#> 132 227.5 185.0 survived
#> 133 227.5 202.5 survived
#> 134 187.5 240.0 survived
#> 135 187.5 240.0 survived
#> 136  97.5 155.0 survived
#> 137  97.5 155.0 survived
#> 138 152.5 182.5 survived
#> 139 152.5 182.5 survived
#> 140 195.0 130.0 survived
#> 141 252.5 150.0 survived
#> 142 135.0 125.0     died
#> 143 135.0 125.0 survived
#> 144 135.0 125.0 survived
#> 145 135.0 125.0 survived
#> 146  72.5 152.5 survived
#> 147  72.5 152.5 survived
#> 148 255.0 230.0 survived
#> 149 255.0 230.0     died
#> 150 255.0 230.0 survived
#> 151 132.5  80.0 survived
#> 152 132.5  80.0 survived
#> 153 132.5  80.0 survived
#> 154 132.5  80.0 survived
#> 155 182.5  55.0 survived
#> 156 182.5  55.0 survived
#> 157 182.5  55.0 survived
#> 158 182.5  55.0 survived
#> 159 280.0 192.5 survived
#> 160 165.0 192.5 survived
#> 161 255.0 230.0 survived
#> 162 280.0 157.5 survived
#> 163 280.0 157.5     died
#> 164 257.5 180.0 survived
#> 165 257.5 180.0 survived
#> 166 132.5  80.0 survived
#> 167 175.0 140.0 survived
#> 168 267.5 127.5 survived
#> 169 132.5  80.0 survived
#> 170 132.5  80.0 survived
#> 171 132.5  80.0 survived
#> 172 185.0 175.0 survived
#> 173 212.5 107.5 survived
#> 174 212.5 107.5 survived
#> 175 212.5 107.5 survived
#> 176  72.5 152.5 survived
#> 177 182.5 200.0 survived
#> 178 182.5 200.0 survived
#> 179 182.5 200.0 survived
#> 180 182.5 200.0 survived
#> 181 182.5 200.0 survived
#> 182 182.5 200.0     died
#> 183 182.5 200.0 survived
#> 184 182.5 200.0 survived
#> 185 205.0 182.5 survived
#> 186 212.5  90.0 survived
#> 187 205.0 182.5 survived
#> 188 212.5  90.0 survived
#> 
#> // tags: gender:gender, outcome:result