Extract a data.frame of all tagged variables

This function returns a data.frame of all the tagged variables stored in a linelist. Note that the output is no longer a linelist, but a regular data.frame.

Usage

tags_df(x)

Arguments

x

a linelist object

Value

A data.frame of tagged variables.

Examples

if (require(outbreaks) && require(magrittr)) {
  ## create a tibble linelist
  x <- measles_hagelloch_1861 %>%
    make_linelist(
      id = "case_ID",
      date_onset = "date_of_prodrome",
      age = "age",
      gender = "gender"
    )
  x

  ## get a data.frame of all tagged variables
  tags_df(x)
}
#>      id date_onset gender age
#> 1     1 1861-11-21      f   7
#> 2     2 1861-11-23      f   6
#> 3     3 1861-11-28      f   4
#> 4     4 1861-11-27      m  13
#> 5     5 1861-11-22      f   8
#> 6     6 1861-11-26      m  12
#> 7     7 1861-11-24      m   6
#> 8     8 1861-11-21      m  10
#> 9     9 1861-11-26      m  13
#> 10   10 1861-11-21      f   7
#> 11   11 1861-11-25      f  11
#> 12   12 1861-11-20      f   7
#> 13   13 1861-11-30      m  13
#> 14   14 1861-11-22      f  13
#> 15   15 1861-11-24      m   8
#> 16   16 1861-11-21      f  15
#> 17   17 1861-11-20      f  10
#> 18   18 1861-11-23      f   2
#> 19   19 1861-11-20      m  11
#> 20   20 1861-11-22      m  10
#> 21   21 1861-11-23      f  13
#> 22   22 1861-11-21      f  10
#> 23   23 1861-11-21      f   7
#> 24   24 1861-11-30      m   4
#> 25   25 1861-11-30      f  12
#> 26   26 1861-11-22      m   7
#> 27   27 1861-11-30      m   5
#> 28   28 1861-11-25      f  10
#> 29   29 1861-11-30      m  13
#> 30   30 1861-11-25      f  11
#> 31   31 1861-11-21      f   9
#> 32   32 1861-11-21      m   7
#> 33   33 1861-11-24      f   7
#> 34   34 1861-11-21      m  11
#> 35   35 1861-11-25      f  13
#> 36   36 1861-12-01      m  11
#> 37   37 1861-11-25      m  13
#> 38   38 1861-11-29   <NA>  12
#> 39   39 1861-11-23      m  10
#> 40   40 1861-12-03      m  13
#> 41   41 1861-11-27      m  12
#> 42   42 1861-11-15      f   4
#> 43   43 1861-11-26      m   2
#> 44   44 1861-11-22      m  10
#> 45   45 1861-11-11      m   7
#> 46   46 1861-11-21      m  13
#> 47   47 1861-11-24      f  11
#> 48   48 1861-12-01      f   3
#> 49   49 1861-11-28      m  10
#> 50   50 1861-12-02      f   6
#> 51   51 1861-12-01      m   4
#> 52   52 1861-11-29      m  13
#> 53   53 1861-11-30      m   6
#> 54   54 1861-12-02      m   4
#> 55   55 1861-12-01      f  11
#> 56   56 1861-11-22      m   8
#> 57   57 1861-12-01      m   3
#> 58   58 1861-11-22      m   9
#> 59   59 1861-11-30      f  10
#> 60   60 1861-12-02      m   2
#> 61   61 1861-12-01      f   5
#> 62   62 1861-12-02      m  14
#> 63   63 1861-11-29      m  12
#> 64   64 1861-11-21      m   7
#> 65   65 1861-12-01      m   2
#> 66   66 1861-12-03      f   5
#> 67   67 1861-12-02      f  11
#> 68   68 1861-11-30      f   2
#> 69   69 1861-12-13      m   1
#> 70   70 1861-11-28      m  13
#> 71   71 1861-12-01      f  10
#> 72   72 1861-12-01      f  10
#> 73   73 1861-11-27      f  11
#> 74   74 1861-11-21      f  10
#> 75   75 1861-11-29      m  13
#> 76   76 1861-12-03      m   2
#> 77   77 1861-11-27      f   8
#> 78   78 1861-11-22      f  11
#> 79   79 1861-11-27      f   5
#> 80   80 1861-11-30      m  12
#> 81   81 1861-12-01      m  12
#> 82   82 1861-11-22      m   8
#> 83   83 1861-12-01      f  10
#> 84   84 1861-11-30      m   6
#> 85   85 1861-12-02      f   5
#> 86   86 1861-12-03      f   3
#> 87   87 1861-12-02      f  12
#> 88   88 1861-12-02      f  10
#> 89   89 1861-12-02      f   3
#> 90   90 1861-12-02      m  11
#> 91   91 1861-12-03      f   4
#> 92   92 1861-12-04      f   2
#> 93   93 1861-11-22      m   8
#> 94   94 1861-11-20      f   4
#> 95   95 1861-12-04      m   1
#> 96   96 1861-11-30      m   2
#> 97   97 1861-11-21      m  10
#> 98   98 1861-12-01      m   3
#> 99   99 1861-12-04      m   5
#> 100 100 1861-12-03      m  12
#> 101 101 1861-12-04      f   7
#> 102 102 1861-12-03      m  12
#> 103 103 1861-12-03   <NA>  12
#> 104 104 1861-12-04      m   5
#> 105 105 1861-12-05      m   3
#> 106 106 1861-11-23      m   4
#> 107 107 1861-11-30      f  12
#> 108 108 1861-12-04      m   6
#> 109 109 1861-12-05      f   6
#> 110 110 1861-12-04      m   3
#> 111 111 1861-12-01      m  12
#> 112 112 1861-12-03      m  10
#> 113 113 1861-12-04      f   0
#> 114 114 1861-12-01   <NA>  13
#> 115 115 1861-12-01      f  11
#> 116 116 1861-11-21      m   8
#> 117 117 1861-12-02      m  14
#> 118 118 1861-12-02      f   2
#> 119 119 1861-12-01      f   0
#> 120 120 1861-12-06      m   1
#> 121 121 1861-12-06      m  10
#> 122 122 1861-12-05      f   1
#> 123 123 1861-12-03      f   1
#> 124 124 1861-12-05      m   3
#> 125 125 1861-12-07      f   2
#> 126 126 1861-12-05   <NA>   5
#> 127 127 1861-12-07      m   1
#> 128 128 1861-12-03      m   5
#> 129 129 1861-12-05      f   4
#> 130 130 1861-11-30      f  12
#> 131 131 1861-12-05      m   1
#> 132 132 1861-12-03      m  11
#> 133 133 1861-12-05      f   2
#> 134 134 1861-12-01      m  13
#> 135 135 1861-12-06      m   2
#> 136 136 1861-11-29      f  13
#> 137 137 1861-12-01      f  10
#> 138 138 1861-12-01   <NA>  11
#> 139 139 1861-12-07   <NA>  13
#> 140 140 1861-12-05   <NA>   2
#> 141 141 1862-01-24      f   4
#> 142 142 1861-12-05      m   5
#> 143 143 1861-12-05      f  11
#> 144 144 1861-12-06      m   2
#> 145 145 1861-11-23      f   8
#> 146 146 1861-12-01      f   4
#> 147 147 1861-12-03      f   0
#> 148 148 1861-12-03      f  13
#> 149 149 1861-12-05      m   4
#> 150 150 1861-12-11      m   0
#> 151 151 1861-12-07      f   2
#> 152 152 1861-12-06      f   4
#> 153 153 1861-11-24      f  10
#> 154 154 1861-12-03      f   6
#> 155 155 1861-12-09      m  13
#> 156 156 1861-11-22      m   8
#> 157 157 1861-12-04      f   4
#> 158 158 1861-12-04      f   3
#> 159 159 1861-12-09      f   2
#> 160 160 1861-12-12      f   0
#> 161 161 1861-12-12      f   6
#> 162 162 1861-12-12      f   6
#> 163 163 1861-12-13      f   1
#> 164 164 1861-12-13      m   3
#> 165 165 1861-12-15      m   2
#> 166 166 1861-12-04      m   1
#> 167 167 1861-12-14      m   0
#> 168 168 1861-12-09      m   1
#> 169 169 1861-12-04      m   4
#> 170 170 1861-12-07      m  10
#> 171 171 1861-12-15      m   0
#> 172 172 1861-11-19      f   3
#> 173 173 1861-11-01      m   6
#> 174 174 1861-11-08      f   3
#> 175 175 1861-11-12      m   2
#> 176 176 1861-12-11   <NA>   0
#> 177 177 1861-11-07      f   8
#> 178 178 1861-11-07      m   4
#> 179 179 1861-11-18      f   1
#> 180 180 1861-11-17      m  10
#> 181 181 1861-11-13      f  10
#> 182 182 1861-11-15      m  13
#> 183 183 1861-11-11      m   4
#> 184 184 1861-10-30   <NA>  13
#> 185 185 1861-12-03      m   3
#> 186 186 1861-11-22   <NA>   6
#> 187 187 1861-12-07      m   0
#> 188 188 1861-11-23   <NA>   1