Check tagged variables are the right class

This function checks the class of each tagged variable in a linelist against pre-defined accepted classes in tags_types().

Usage

validate_types(x, ref_types = tags_types())

Arguments

x

a linelist object

ref_types

a list providing allowed types for all tags, as returned by tags_types()

Value

A named list.

See also

Examples

if (require(outbreaks) && require(magrittr)) {
  ## create an invalid linelist - gender is a numeric
  x <- measles_hagelloch_1861 %>%
    make_linelist(
      id = "case_ID",
      gender = "infector"
    )
  x

  ## the below would issue an error
  ## note: tryCatch is only used to avoid a genuine error in the example
  tryCatch(validate_types(x), error = paste)

  ## to allow other types, e.g. gender to be integer, character or factor
  validate_types(x, tags_types(gender = c("integer", "character", "factor")))
}
#> 
#> // linelist object
#>     case_ID infector date_of_prodrome date_of_rash date_of_death age gender
#> 1         1       45       1861-11-21   1861-11-25          <NA>   7      f
#> 2         2       45       1861-11-23   1861-11-27          <NA>   6      f
#> 3         3      172       1861-11-28   1861-12-02          <NA>   4      f
#> 4         4      180       1861-11-27   1861-11-28          <NA>  13      m
#> 5         5       45       1861-11-22   1861-11-27          <NA>   8      f
#> 6         6      180       1861-11-26   1861-11-29          <NA>  12      m
#> 7         7       42       1861-11-24   1861-11-28          <NA>   6      m
#> 8         8       45       1861-11-21   1861-11-26          <NA>  10      m
#> 9         9      182       1861-11-26   1861-11-30          <NA>  13      m
#> 10       10       45       1861-11-21   1861-11-25          <NA>   7      f
#> 11       11      182       1861-11-25   1861-11-30          <NA>  11      f
#> 12       12       45       1861-11-20   1861-11-25          <NA>   7      f
#> 13       13       12       1861-11-30   1861-12-05          <NA>  13      m
#> 14       14      181       1861-11-22   1861-11-29          <NA>  13      f
#> 15       15       45       1861-11-24   1861-11-29          <NA>   8      m
#> 16       16      181       1861-11-21   1861-11-25          <NA>  15      f
#> 17       17      181       1861-11-20   1861-11-25          <NA>  10      f
#> 18       18      175       1861-11-23   1861-11-27          <NA>   2      f
#> 19       19      181       1861-11-20   1861-11-24          <NA>  11      m
#> 20       20      181       1861-11-22   1861-11-27          <NA>  10      m
#> 21       21      181       1861-11-23   1861-11-29          <NA>  13      f
#> 22       22       45       1861-11-21   1861-11-25          <NA>  10      f
#> 23       23       45       1861-11-21   1861-11-25          <NA>   7      f
#> 24       24       22       1861-11-30   1861-12-04          <NA>   4      m
#> 25       25       22       1861-11-30   1861-12-04          <NA>  12      f
#> 26       26       45       1861-11-22   1861-11-27          <NA>   7      m
#> 27       27       10       1861-11-30   1861-12-02    1861-12-14   5      m
#> 28       28      180       1861-11-25   1861-11-30          <NA>  10      f
#> 29       29       31       1861-11-30   1861-12-02          <NA>  13      m
#> 30       30       45       1861-11-25   1861-11-29          <NA>  11      f
#> 31       31       45       1861-11-21   1861-11-24          <NA>   9      f
#> 32       32       45       1861-11-21   1861-11-24          <NA>   7      m
#> 33       33       45       1861-11-24   1861-11-26          <NA>   7      f
#> 34       34      181       1861-11-21   1861-11-26          <NA>  11      m
#> 35       35      182       1861-11-25   1861-11-30          <NA>  13      f
#> 36       36       34       1861-12-01   1861-12-05          <NA>  11      m
#> 37       37      182       1861-11-25   1861-12-02          <NA>  13      m
#> 38       38       17       1861-11-29   1861-12-05          <NA>  12   <NA>
#> 39       39       45       1861-11-23   1861-11-26          <NA>  10      m
#> 40       40       93       1861-12-03   1861-12-07          <NA>  13      m
#> 41       41      180       1861-11-27   1861-11-29          <NA>  12      m
#> 42       42      178       1861-11-15   1861-11-19          <NA>   4      f
#> 43       43       42       1861-11-26   1861-11-30          <NA>   2      m
#> 44       44       45       1861-11-22   1861-11-24          <NA>  10      m
#> 45       45      184       1861-11-11   1861-11-13          <NA>   7      m
#> 46       46       45       1861-11-21   1861-11-27          <NA>  13      m
#> 47       47       45       1861-11-24   1861-11-27          <NA>  11      f
#> 48       48       10       1861-12-01   1861-12-05          <NA>   3      f
#> 49       49       17       1861-11-28   1861-12-06          <NA>  10      m
#> 50       50        8       1861-12-02   1861-12-06          <NA>   6      f
#> 51       51       31       1861-12-01   1861-12-05          <NA>   4      m
#> 52       52       17       1861-11-29   1861-12-05    1861-12-12  13      m
#> 53       53       17       1861-11-30   1861-12-04          <NA>   6      m
#> 54       54       17       1861-12-02   1861-12-05    1861-12-15   4      m
#> 55       55       17       1861-12-01   1861-12-05          <NA>  11      f
#> 56       56       45       1861-11-22   1861-11-29          <NA>   8      m
#> 57       57       56       1861-12-01   1861-12-05          <NA>   3      m
#> 58       58       45       1861-11-22   1861-11-26          <NA>   9      m
#> 59       59       58       1861-11-30   1861-12-04          <NA>  10      f
#> 60       60       58       1861-12-02   1861-12-06    1861-12-12   2      m
#> 61       61      186       1861-12-01   1861-12-03          <NA>   5      f
#> 62       62       11       1861-12-02   1861-12-06          <NA>  14      m
#> 63       63       19       1861-11-29   1861-12-08          <NA>  12      m
#> 64       64       45       1861-11-21   1861-11-27          <NA>   7      m
#> 65       65       64       1861-12-01   1861-12-06          <NA>   2      m
#> 66       66       64       1861-12-03   1861-12-05          <NA>   5      f
#> 67       67       11       1861-12-02   1861-12-07          <NA>  11      f
#> 68       68      179       1861-11-30   1861-12-02          <NA>   2      f
#> 69       69       54       1861-12-13   1861-12-17          <NA>   1      m
#> 70       70      180       1861-11-28   1861-12-03          <NA>  13      m
#> 71       71       10       1861-12-01   1861-12-05          <NA>  10      f
#> 72       72       12       1861-12-01   1861-12-06          <NA>  10      f
#> 73       73      180       1861-11-27   1861-12-05          <NA>  11      f
#> 74       74       45       1861-11-21   1861-11-26          <NA>  10      f
#> 75       75       74       1861-11-29   1861-12-05          <NA>  13      m
#> 76       76        5       1861-12-03   1861-12-06          <NA>   2      m
#> 77       77      180       1861-11-27   1861-11-27          <NA>   8      f
#> 78       78      181       1861-11-22   1861-11-26          <NA>  11      f
#> 79       79      179       1861-11-27   1861-12-05          <NA>   5      f
#> 80       80       78       1861-11-30   1861-12-05    1861-12-12  12      m
#> 81       81       39       1861-12-01   1861-12-04          <NA>  12      m
#> 82       82       45       1861-11-22   1861-11-26          <NA>   8      m
#> 83       83       82       1861-12-01   1861-12-05          <NA>  10      f
#> 84       84       82       1861-11-30   1861-12-04          <NA>   6      m
#> 85       85       44       1861-12-02   1861-12-05          <NA>   5      f
#> 86       86        1       1861-12-03   1861-12-05          <NA>   3      f
#> 87       87       47       1861-12-02   1861-12-07          <NA>  12      f
#> 88       88       47       1861-12-02   1861-12-07          <NA>  10      f
#> 89       89       12       1861-12-02   1861-12-05          <NA>   3      f
#> 90       90       93       1861-12-02   1861-12-05          <NA>  11      m
#> 91       91       93       1861-12-03   1861-12-05          <NA>   4      f
#> 92       92       93       1861-12-04   1861-12-07          <NA>   2      f
#> 93       93       45       1861-11-22   1861-11-26          <NA>   8      m
#> 94       94      183       1861-11-20   1861-12-03          <NA>   4      f
#> 95       95       10       1861-12-04   1861-12-07          <NA>   1      m
#> 96       96       97       1861-11-30   1861-12-04          <NA>   2      m
#> 97       97       45       1861-11-21   1861-11-25          <NA>  10      m
#> 98       98       64       1861-12-01   1861-12-12    1861-12-17   3      m
#> 99       99       11       1861-12-04   1861-12-10          <NA>   5      m
#> 100     100       47       1861-12-03   1861-12-07          <NA>  12      m
#> 101     101        7       1861-12-04   1861-12-06          <NA>   7      f
#> 102     102       21       1861-12-03   1861-12-07          <NA>  12      m
#> 103     103       37       1861-12-03   1861-12-07          <NA>  12   <NA>
#> 104     104       58       1861-12-04   1861-12-06          <NA>   5      m
#> 105     105       74       1861-12-05   1861-12-08          <NA>   3      m
#> 106     106       42       1861-11-23   1861-11-26          <NA>   4      m
#> 107     107       19       1861-11-30   1861-12-07          <NA>  12      f
#> 108     108      106       1861-12-04   1861-12-06          <NA>   6      m
#> 109     109       12       1861-12-05   1861-12-07          <NA>   6      f
#> 110     110       18       1861-12-04   1861-12-06          <NA>   3      m
#> 111     111       34       1861-12-01   1861-12-05          <NA>  12      m
#> 112     112       21       1861-12-03   1861-12-08          <NA>  10      m
#> 113     113       31       1861-12-04   1861-12-07          <NA>   0      f
#> 114     114       78       1861-12-01   1861-12-05          <NA>  13   <NA>
#> 115     115       16       1861-12-01   1861-12-07          <NA>  11      f
#> 116     116       45       1861-11-21   1861-11-25          <NA>   8      m
#> 117     117      116       1861-12-02   1861-12-06          <NA>  14      m
#> 118     118      116       1861-12-02   1861-12-06          <NA>   2      f
#> 119     119      116       1861-12-01   1861-12-08          <NA>   0      f
#> 120     120        7       1861-12-06   1861-12-08          <NA>   1      m
#> 121     121       11       1861-12-06   1861-12-08          <NA>  10      m
#> 122     122      188       1861-12-05   1861-12-10          <NA>   1      f
#> 123     123        7       1861-12-03   1861-12-07    1861-12-13   1      f
#> 124     124        7       1861-12-05   1861-12-08          <NA>   3      m
#> 125     125        7       1861-12-07   1861-12-09          <NA>   2      f
#> 126     126       37       1861-12-05   1861-12-09          <NA>   5   <NA>
#> 127     127      106       1861-12-07   1861-12-09    1861-12-15   1      m
#> 128     128        7       1861-12-03   1861-12-05          <NA>   5      m
#> 129     129        7       1861-12-05   1861-12-07          <NA>   4      f
#> 130     130       56       1861-11-30   1861-12-06          <NA>  12      f
#> 131     131       56       1861-12-05   1861-12-06          <NA>   1      m
#> 132     132       14       1861-12-03   1861-12-07          <NA>  11      m
#> 133     133       18       1861-12-05   1861-12-08          <NA>   2      f
#> 134     134       78       1861-12-01   1861-12-06          <NA>  13      m
#> 135     135       79       1861-12-06   1861-12-07          <NA>   2      m
#> 136     136       17       1861-11-29   1861-12-07          <NA>  13      f
#> 137     137       16       1861-12-01   1861-12-06          <NA>  10      f
#> 138     138       34       1861-12-01   1861-12-05          <NA>  11   <NA>
#> 139     139        4       1861-12-07   1861-12-11          <NA>  13   <NA>
#> 140     140        6       1861-12-05   1861-12-10          <NA>   2   <NA>
#> 141     141       NA       1862-01-24   1862-01-27          <NA>   4      f
#> 142     142      145       1861-12-05   1861-12-08    1861-12-09   5      m
#> 143     143      145       1861-12-05   1861-12-08          <NA>  11      f
#> 144     144      145       1861-12-06   1861-12-08          <NA>   2      m
#> 145     145       45       1861-11-23   1861-11-27          <NA>   8      f
#> 146     146      172       1861-12-01   1861-12-07          <NA>   4      f
#> 147     147       18       1861-12-03   1861-12-07          <NA>   0      f
#> 148     148       14       1861-12-03   1861-12-07          <NA>  13      f
#> 149     149       39       1861-12-05   1861-12-09    1861-12-15   4      m
#> 150     150      148       1861-12-11   1861-12-15          <NA>   0      m
#> 151     151      153       1861-12-07   1861-12-11          <NA>   2      f
#> 152     152      153       1861-12-06   1861-12-08          <NA>   4      f
#> 153     153       45       1861-11-24   1861-11-27          <NA>  10      f
#> 154     154      153       1861-12-03   1861-12-04          <NA>   6      f
#> 155     155       73       1861-12-09   1861-12-13          <NA>  13      m
#> 156     156       45       1861-11-22   1861-11-24          <NA>   8      m
#> 157     157      156       1861-12-04   1861-12-07          <NA>   4      f
#> 158     158      156       1861-12-04   1861-12-07          <NA>   3      f
#> 159     159       37       1861-12-09   1861-12-13          <NA>   2      f
#> 160     160       68       1861-12-12   1861-12-13          <NA>   0      f
#> 161     161      148       1861-12-12   1861-12-16          <NA>   6      f
#> 162     162      123       1861-12-12   1861-12-16          <NA>   6      f
#> 163     163      123       1861-12-13   1861-12-16    1861-12-28   1      f
#> 164     164      102       1861-12-13   1861-12-15          <NA>   3      m
#> 165     165      102       1861-12-15   1861-12-19          <NA>   2      m
#> 166     166      153       1861-12-04   1861-12-07          <NA>   1      m
#> 167     167      110       1861-12-14   1861-12-18          <NA>   0      m
#> 168     168       98       1861-12-09   1861-12-15          <NA>   1      m
#> 169     169      153       1861-12-04   1861-12-07          <NA>   4      m
#> 170     170      153       1861-12-07   1861-12-10          <NA>  10      m
#> 171     171      169       1861-12-15   1861-12-17          <NA>   0      m
#> 172     172      174       1861-11-19   1861-11-21          <NA>   3      f
#> 173     173       NA       1861-11-01   1861-11-03          <NA>   6      m
#> 174     174       NA       1861-11-08   1861-11-08          <NA>   3      f
#> 175     175      173       1861-11-12   1861-11-15          <NA>   2      m
#> 176     176      146       1861-12-11   1861-12-15          <NA>   0   <NA>
#> 177     177      184       1861-11-07   1861-11-11          <NA>   8      f
#> 178     178      184       1861-11-07   1861-11-11          <NA>   4      m
#> 179     179      177       1861-11-18   1861-11-21          <NA>   1      f
#> 180     180      177       1861-11-17   1861-11-21          <NA>  10      m
#> 181     181      184       1861-11-13   1861-11-17          <NA>  10      f
#> 182     182      184       1861-11-15   1861-11-18    1861-11-18  13      m
#> 183     183      184       1861-11-11   1861-11-15          <NA>   4      m
#> 184     184       NA       1861-10-30   1861-11-06          <NA>  13   <NA>
#> 185     185       82       1861-12-03   1861-12-07          <NA>   3      m
#> 186     186       45       1861-11-22   1861-11-26          <NA>   6   <NA>
#> 187     187       82       1861-12-07   1861-12-11          <NA>   0      m
#> 188     188      175       1861-11-23   1861-11-27          <NA>   1   <NA>
#>     family_ID class complications x_loc y_loc
#> 1          41     1           yes 142.5 100.0
#> 2          41     1           yes 142.5 100.0
#> 3          41     0           yes 142.5 100.0
#> 4          61     2           yes 165.0 102.5
#> 5          42     1           yes 145.0 120.0
#> 6          42     2           yes 145.0 120.0
#> 7          26     0           yes 272.5 147.5
#> 8          44     1           yes  97.5 155.0
#> 9          44     2           yes  97.5 155.0
#> 10         29     1           yes 240.0  75.0
#> 11         27     2           yes 270.0 135.0
#> 12         32     1           yes 195.0  27.5
#> 13         32     2           yes 195.0  27.5
#> 14         22     2           yes 227.5 185.0
#> 15         22     1           yes 227.5 185.0
#> 16         43     2           yes 172.5 172.5
#> 17         43     2           yes 172.5 172.5
#> 18         43     0           yes 172.5 172.5
#> 19         11     2           yes 167.5   5.0
#> 20         11     2           yes 167.5   5.0
#> 21         11     2           yes 167.5   5.0
#> 22         35     1           yes 167.5   5.0
#> 23         35     1           yes 167.5   5.0
#> 24         35     0           yes 167.5   5.0
#> 25         35     2           yes 167.5   5.0
#> 26         67     1           yes   7.5  37.5
#> 27         29     0           yes 240.0  75.0
#> 28         65     2           yes  15.0  47.5
#> 29         15     2           yes 125.0 187.5
#> 30         15     2           yes 125.0 187.5
#> 31         15     1           yes 125.0 187.5
#> 32         15     1           yes 125.0 187.5
#> 33          1     1           yes 125.0 187.5
#> 34          1     2           yes 125.0 187.5
#> 35         10     2           yes 190.0 115.0
#> 36         46     2           yes 280.0 192.5
#> 37         46     2           yes 280.0 192.5
#> 38         54     2           yes 280.0 192.5
#> 39          5     1           yes 227.5 217.5
#> 40         36     2           yes 162.5  47.5
#> 41         59     2           yes 185.0  80.0
#> 42         33     0           yes 170.0  17.5
#> 43         33     0           yes 170.0  17.5
#> 44         45     1           yes 150.0  15.0
#> 45         48     1           yes  75.0  55.0
#> 46         48     2           yes  75.0  55.0
#> 47         48     2           yes  75.0  55.0
#> 48         29     0           yes 240.0  75.0
#> 49         49     2           yes 175.0 140.0
#> 50         44     0           yes  97.5 155.0
#> 51         15     0           yes 125.0 187.5
#> 52         43     2           yes 172.5 172.5
#> 53         43     0           yes 172.5 172.5
#> 54         43     0           yes 172.5 172.5
#> 55         43     2           yes 172.5 172.5
#> 56         20     1           yes 265.0 225.0
#> 57         20     0           yes 265.0 225.0
#> 58         24     1           yes 250.0 210.0
#> 59         24     2           yes 250.0 210.0
#> 60         24     0           yes 250.0 210.0
#> 61          8     0           yes 270.0 102.5
#> 62          8     2           yes 270.0 102.5
#> 63          8     2           yes 270.0 102.5
#> 64         30     1           yes 237.5  90.0
#> 65         30     0           yes 237.5  90.0
#> 66         30     0           yes 237.5  90.0
#> 67         55     2           yes 267.5 127.5
#> 68         16     0           yes 165.0 192.5
#> 69         63     0           yes 150.0 165.0
#> 70         21     2           yes 205.0 182.5
#> 71         29     2           yes 240.0  75.0
#> 72         32     2           yes 195.0  27.5
#> 73         47     2           yes 182.5  55.0
#> 74         23     1           yes 257.5 195.0
#> 75         23     2           yes 257.5 195.0
#> 76         42     0           yes 145.0 120.0
#> 77         18     1           yes 187.5 240.0
#> 78         18     2           yes 187.5 240.0
#> 79         18     0           yes 187.5 240.0
#> 80         18     2           yes 187.5 240.0
#> 81          5     2           yes 227.5 217.5
#> 82         21     1           yes 205.0 182.5
#> 83         21     2           yes 205.0 182.5
#> 84         21     0           yes 205.0 182.5
#> 85         45     0           yes 150.0  15.0
#> 86         41     0           yes 142.5 100.0
#> 87         60     2           yes  77.5  42.5
#> 88         60     2           yes  77.5  42.5
#> 89         32     0           yes 195.0  27.5
#> 90         36     2           yes 162.5  47.5
#> 91         36     0           yes 162.5  47.5
#> 92         36     0           yes 162.5  47.5
#> 93         36     1           yes 162.5  47.5
#> 94         10     0           yes 190.0 115.0
#> 95         29     0           yes 240.0  75.0
#> 96         28     0           yes 230.0 120.0
#> 97         28     1           yes 230.0 120.0
#> 98         27     0           yes 270.0 135.0
#> 99         27     0           yes 270.0 135.0
#> 100        69     2           yes 252.5 150.0
#> 101        26     0           yes 272.5 147.5
#> 102        68     2           yes 257.5 180.0
#> 103        54     2           yes 280.0 192.5
#> 104        24     0           yes 250.0 210.0
#> 105        23     0           yes 257.5 195.0
#> 106        34     0           yes 170.0  17.5
#> 107        34     2           yes 170.0  17.5
#> 108        34     0           yes 170.0  17.5
#> 109        32     0           yes 195.0  27.5
#> 110        49     0           yes 175.0 140.0
#> 111         2     2           yes 142.5 180.0
#> 112         2     2           yes 142.5 180.0
#> 113        15     0           yes 125.0 187.5
#> 114        52     2           yes 200.0 210.0
#> 115        66     2           yes  75.0  20.0
#> 116        40     1           yes 127.5 147.5
#> 117        40     2           yes 127.5 147.5
#> 118        40     0           yes 127.5 147.5
#> 119        40     0           yes 127.5 147.5
#> 120        26     0           yes 272.5 147.5
#> 121        50     2           yes 252.5 150.0
#> 122        50     0           yes 252.5 150.0
#> 123        25     0           yes 280.0 157.5
#> 124         7     0           yes 280.0 167.5
#> 125         7     0           yes 280.0 167.5
#> 126        54     0           yes 280.0 192.5
#> 127        34     0           yes 170.0  17.5
#> 128         6     0           yes 240.0 225.0
#> 129         6     0           yes 240.0 225.0
#> 130        20     2           yes 265.0 225.0
#> 131        20     0           yes 265.0 225.0
#> 132        22     2           yes 227.5 185.0
#> 133        53     0           yes 227.5 202.5
#> 134        18     2           yes 187.5 240.0
#> 135        18     0           yes 187.5 240.0
#> 136        12     2           yes  97.5 155.0
#> 137        12     2           yes  97.5 155.0
#> 138         3     2           yes 152.5 182.5
#> 139         3     2           yes 152.5 182.5
#> 140        58     0           yes 195.0 130.0
#> 141        50     0           yes 252.5 150.0
#> 142        39     0           yes 135.0 125.0
#> 143        39     2           yes 135.0 125.0
#> 144        39     0           yes 135.0 125.0
#> 145        39     1           yes 135.0 125.0
#> 146        13     0           yes  72.5 152.5
#> 147        13     0           yes  72.5 152.5
#> 148        19     2           yes 255.0 230.0
#> 149        19     0           yes 255.0 230.0
#> 150        19     0           yes 255.0 230.0
#> 151        37     0           yes 132.5  80.0
#> 152        37     0           yes 132.5  80.0
#> 153        37     1           yes 132.5  80.0
#> 154        37     0           yes 132.5  80.0
#> 155        47     2           yes 182.5  55.0
#> 156        31     1           yes 182.5  55.0
#> 157        31     0           yes 182.5  55.0
#> 158        31     0           yes 182.5  55.0
#> 159        46     0           yes 280.0 192.5
#> 160        16     0           yes 165.0 192.5
#> 161        19     0           yes 255.0 230.0
#> 162        25     0           yes 280.0 157.5
#> 163        25     0           yes 280.0 157.5
#> 164        14     0           yes 257.5 180.0
#> 165        14     0           yes 257.5 180.0
#> 166        37     0           yes 132.5  80.0
#> 167        49     0           yes 175.0 140.0
#> 168        56     0           yes 267.5 127.5
#> 169        38     0           yes 132.5  80.0
#> 170        38     2           yes 132.5  80.0
#> 171        38     0           yes 132.5  80.0
#> 172        62     0           yes 185.0 175.0
#> 173         9     0           yes 212.5 107.5
#> 174         9     0           yes 212.5 107.5
#> 175         9     0           yes 212.5 107.5
#> 176        64     0           yes  72.5 152.5
#> 177        17     1           yes 182.5 200.0
#> 178        17     0           yes 182.5 200.0
#> 179        17     1           yes 182.5 200.0
#> 180        17     2           yes 182.5 200.0
#> 181         4     2           yes 182.5 200.0
#> 182         4     2           yes 182.5 200.0
#> 183         4     0           yes 182.5 200.0
#> 184        51     2           yes 182.5 200.0
#> 185        21     0           yes 205.0 182.5
#> 186        57     0           yes 212.5  90.0
#> 187        21     0           yes 205.0 182.5
#> 188        57     0           yes 212.5  90.0
#> 
#> // tags: id:case_ID, gender:infector